refactor(components): [checkbox]split composables into new file
