+++ packages/components/checkbox/index.ts
@@ -15,4 +15,3 @@ export const ElCheckboxGroup = withNoopInstall(CheckboxGroup)
 
 export * from './src/checkbox-group'
 export * from './src/checkbox'
-export * from './src/checkbox.type'
