+++ packages/components/checkbox/src/checkbox.vue
@@ -66,7 +66,8 @@
 <script lang="ts" setup>
 import { useSlots } from 'vue'
 import { useNamespace } from '@element-plus/hooks'
-import { checkboxEmits, checkboxProps, useCheckbox } from './checkbox'
+import { checkboxEmits, checkboxProps } from './checkbox'
+import { useCheckbox } from './composables'
 
 defineOptions({
   name: 'ElCheckbox',
